# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Farah API Documentation",
        "description": "توثيق واجهات برمجة التطبيقات (APIs) لمنصة فرح لتطبيق Flutter",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://farahbackend-production.up.railway.app",
            "description": "سيرفر الإنتاج (Railway Production)"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "السيرفر المحلي (Local Server)"
        }
    ],
    "paths": {
...
        "/api/services/{id}": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "عرض تفاصيل خدمة محددة",
                "operationId": "dce5b306a141bdec91eff8a15534d7d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "معرّف الخدمة",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم جلب تفاصيل الخدمة بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "category_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "provider_id": {
                                                    "type": "integer",
                                                    "example": 4
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "خدمة تصوير وتوثيق الحفلات"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "تغطية شاملة وفيديو بدقة عالية"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 250
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "active"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2026-08-26T12:00:00.000000Z"
                                                },
                                                "category": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "تصوير وفيديو"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "provider": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 4
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "أحمد مصور"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "ahmed@example.com"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "images": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "image_path": {
                                                                "type": "string",
                                                                "example": "services/image1.jpg"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "reviews": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "rating": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "comment": {
                                                                "type": "string",
                                                                "example": "خدمة ممتازة جداً"
                                                            },
                                                            "user": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer",
                                                                        "example": 8
                                                                    },
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "محمد علي"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "الخدمة غير موجودة"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "أدخل التوكن بهذا الشكل: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}